code/python/yorishbot/requirements.txt
line-bot-sdk
flask
python-dotenv
pyngrok
pyaml